refactor(services): extract services layer with compatibility shims [roadmap:v0.7.4]
Move inspect, improve, relationships, portfolio, stats, ingest, and diff into rac.services. Service modules import core via canonical rac.core.* paths (parser->markdown, validate->validation) and keep service siblings relative. Thin shims remain at the previous top-level paths so existing imports keep resolving; the rac.ingest shim also forwards the private helpers the tests use. No behavior change.
